+++ src/modules/mo-product-dto/product/data-transfer-objects/cluster/index.ts
@@ -1,3 +1,5 @@
+export * from './k8s-manager';
+
 export * from './cluster.dto';
 export * from './cluster-flat.dto';
 export * from './cluster-create-request.dto';