+++ src/modules/mo-project-dto/data-transfer-objects/k8s/k8s-project.dto.ts
@@ -5,8 +5,8 @@ export class K8sProjectDto {
   @Expose()
   id: string;
 
-  @Expose()
-  shortId: string;
+  // @Expose()
+  // shortId: string;
 
   @Expose()
   displayName: string;